Add machine learning modules and required installations
This commit introduces a main.py file which contains four functions for linear and polynomial regression (both simple and multiple). The purpose of these functions is to provide basic instances of modeling and predictions in machine learning. requirements.txt has also been added to ensure all necessary installations are clear for functionality, specifically for libraries such as numpy, sklearn etc. Additionally, a README.md file has been introduced for a basic explanation of the project.